+++ dags/cluster/tasks/business_logic/cluster_acteurs_suggestions/prepare.py
@@ -1,8 +1,9 @@
 import pandas as pd
 from cluster.config.constants import COL_PARENT_DATA_NEW
-from cluster.tasks.business_logic.misc.data_serialize_reconstruct import data_serialize
 from utils import logging_utils as log
 from utils.django import django_setup_full
+
+from dags.utils.data_serialize_reconstruct import data_serialize
 
 django_setup_full()
