+++ dags/cluster/tasks/business_logic/misc/cluster_exclude_intra_source.py
@@ -4,35 +4,27 @@ import pandas as pd
 def cluster_exclude_intra_source(
     df: pd.DataFrame,
 ) -> tuple[pd.DataFrame, pd.DataFrame | None]:
-    """
-    Exclue les acteurs d'une même source (si il y en a) d'un cluster donné,
-    en ne conservant qu'1 seul acteur par source
+    """Exclude intra-source actors from a cluster. Returns 2 dfs:
+    - original without duplication on sources
+    - another df with excluded actors"""
 
-    Args:
-        df (pd.DataFrame): DataFrame contenant les acteurs du cluster
-
-    Returns:
-        tuple[pd.DataFrame, pd.DataFrame]: 1er élément: DataFrame des acteurs
-        conservés, 2ème élément: DataFrame des acteurs exclus
-    """
-    # Pour l'instant on contraint l'utilisation de cette fonction sur 1 seul cluster
+    # Restrict usage of function to single cluster (simplifies logic)
     if not df["cluster_id"].nunique() == 1:
         raise ValueError("Fonction à utiliser sur 1 seul cluster")
 
-    # Les acteurs parents n'ont pas de source_id, et donc ne sont pas concernés
+    # Parents have not source thus ignored
     df_children = df[df["source_id"].notnull()]
     df_parents = df[df["source_id"].isnull()]
 
-    # Pas de doublons de source -> on retourne le df d'origine
+    # No duplication on sources -> return original df as is
     if df_children["source_id"].nunique() == len(df_children):
         return df, None
 
-    # Pour chaque source, on ne garde qu'un seul acteur
-    # (pour l'instant on garde le premier)
-    # TODO: améliorer avec les scores de similarité
+    # For each source, keep only one actor
+    # TODO: could be improved with similarity scores
     df_children_kept = df_children.groupby("source_id").first().reset_index()
 
-    # Résultat final: acteurs conservés vs. exclus
+    # Final results: acteurs kept vs. excluded
     id = "identifiant_unique"
     df_kept = pd.concat([df_children_kept, df_parents], ignore_index=True)
     df_lost = df_children[~df_children[id].isin(df_kept[id])].reset_index(drop=True)
