+++ dags/shared/config/start_dates.py
@@ -6,10 +6,11 @@ we configure start dates to be 1 interval in the past of schedule
 present/future date) thus if there is catchup, it's only 1 run max."""
 
 from dataclasses import dataclass
-from datetime import datetime, timedelta
+from datetime import datetime, timedelta, timezone
+
+NOW = datetime.now(timezone.utc)
 
 
 @dataclass(frozen=True)
 class START_DATES:
-    FOR_SCHEDULE_DAILY: datetime = datetime.now() - timedelta(days=1)
-    FOR_SCHEDULE_NONE: datetime = datetime.now() - timedelta(days=1)
+    YESTERDAY: datetime = NOW - timedelta(days=1)
