+++ dags/tools/dags/airflow_cleanup_db.py
@@ -26,7 +26,7 @@ DAYS_TO_KEEP = 7
     dag_id="airflow_cleanup_db",
     dag_display_name="Maintenance - Airflow - Nettoyer la DB (XCOM, logs, etc.)",
     schedule=SCHEDULES.DAILY,
-    start_date=START_DATES.FOR_SCHEDULE_DAILY,
+    start_date=START_DATES.YESTERDAY,
     catchup=CATCHUPS.AWLAYS_FALSE,
     is_paused_upon_creation=False,
     render_template_as_native_obj=True,
