+++ dags_unit_tests/utils/test_data_serialize_reconstruct.py
@@ -1,12 +1,9 @@
 from datetime import datetime
 
 import pytest
-from cluster.tasks.business_logic.misc.data_serialize_reconstruct import (
-    data_reconstruct,
-    data_serialize,
-)
 from django.contrib.gis.geos import Point
 from rich import print
+from utils.data_serialize_reconstruct import data_reconstruct, data_serialize
 
 from qfdmo.models.acteur import RevisionActeur
 from unit_tests.qfdmo.acteur_factory import (
