+++ data/models/changes/acteur_create_as_parent.py
@@ -4,9 +4,7 @@ change_model to create a parent acteur
 
 """
 
-from dags.cluster.tasks.business_logic.misc.data_serialize_reconstruct import (
-    data_reconstruct,
-)
+from dags.utils.data_serialize_reconstruct import data_reconstruct
 from data.models.changes.acteur_abstract import ChangeActeurAbstract
 from qfdmo.models import ActeurStatus, RevisionActeur
