+++ data/models/changes/acteur_update_data.py
@@ -2,9 +2,7 @@
 about an acteur, taking care of handling Acteur vs. RevisionActeur
 and data reconstruction."""
 
-from dags.cluster.tasks.business_logic.misc.data_serialize_reconstruct import (
-    data_reconstruct,
-)
+from dags.utils.data_serialize_reconstruct import data_reconstruct
 from data.models.changes.acteur_abstract import ChangeActeurAbstract
 from qfdmo.models import Acteur, RevisionActeur
